ManagerPrototype - working on check employee method

+++ ManagerPrototype/src/main/java/fviv/model/EmployeeRepository.java
@@ -1,7 +1,8 @@
 package fviv.model;
 
+
 import org.springframework.data.repository.CrudRepository;
 
 public interface EmployeeRepository extends CrudRepository<Employee, Long>{
-	Employee findById(int id);
+	Employee findById(long id);
 }
